Add Test Framework to implement
For now all the tests are undefined but the framework is there, it is
likely to be used in development for the token analyser and in
production for the global testing of the program.
The cabal file had to be modified by removing error flags for tests
because all the tests are undefined and therefore cannot be run. They
are therefore never called and create a warning that had to be dissabled