Change instructions registers data
Modify the show instances to make them printARM instructions in order
not to overlap the show patterns.
Change the registers to not be a genral register taking an int but the
actual list of registers. It has been made an instance of boundable so
we can call the registers using the bounds instead of a possible int
list